Move To/FromEigen to separate file
