+++ test/mechanics/dofs/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_unit_test(DofContainer)
 add_unit_test(DofVector)
+add_unit_test(DofVectorConvertEigen)
 add_unit_test(DofMatrix)
 add_unit_test(DofMatrixSparse)
 add_unit_test(DofType)
